# src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Search API",
    "description": "Privacy-respecting metasearch engine API",
    "version": "1.0.0",
    "contact": {
      "name": "Search Support"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API v1"
    }
  ],
...
  "paths": {
...
    "/autocomplete": {
      "get": {
        "summary": "Autocomplete suggestions",
        "description": "Get autocomplete suggestions for a partial query",
        "operationId": "autocomplete",
        "tags": ["Search"],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "description": "Partial search query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Autocomplete suggestions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutocompleteResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API token authentication. Use token with prefix: usr_ for user tokens."
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "Session cookie authentication"
      }
    },
    "schemas": {
...
      "AutocompleteResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  }
}